+++ js/main.js
@@ -1,2 +1,3 @@
 import './map.js';
 import './form.js';
+import './api.js';
